Add the Navbar to the project
