+++ src/app/layout.tsx
@@ -1,5 +1,7 @@
+"use client";
 import "./globals.css";
 import { Inter } from "next/font/google";
+import Navbar from "./navbar";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -15,7 +17,10 @@ export default function RootLayout({
 }) {
     return (
         <html lang="en">
-            <body className={inter.className}>{children}</body>
+            <body className={inter.className}>
+                <Navbar />
+                <div className="px-8"> {children}</div>
+            </body>
         </html>
     );
 }
